@@ -1,29 +1,31 @@
 name: steel_challenge_match_tracker
 description: A new Flutter application.
 
-version: 2.2.7+23
+#version: 2.2.7+23
+version: 2.2.8+24
 
 environment:
 # sdk: ">=2.2.0 <3.0.0"
 #  sdk: ">=2.17.0-0 <3.0.0"
-  sdk: ">=2.10.0-0 <3.0.0"
+#  sdk: ">=2.10.0-0 <3.0.0"
+  sdk: ">=3.5.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  shared_preferences: ^2.0.10
-  rflutter_alert: ^2.0.4
-  url_launcher: ^6.0.9
-  sqflite: ^2.0.0+3
-  path_provider: ^2.0.2
-#  just_audio: ^0.9.34
-  just_audio: ^0.9.32
-  font_awesome_flutter: ^10.3.0
+  shared_preferences: ^2.3.3
+  rflutter_alert: ^2.0.7
+  url_launcher: ^6.3.1
+  sqflite: ^2.4.1
+  path_provider: ^2.1.5
+  path: ^1.9.0
+  just_audio: ^0.9.42
+  font_awesome_flutter: ^10.8.0
 
 dev_dependencies:
-  flutter_lints: ^1.0.4
+  flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
 
