@@ -13,7 +13,7 @@ void main() {
 
 /// This Widget is the main application widget.
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
